Refatoração arquitetural para pastas core, utils e reports
Move toda a lógica de roteirização, geocodificação e IO da rota.py
para os submódulos apropriados conforme os novos diretórios de
arquitetura (utils, core e reports), transformando a rota.py em um
orquestrador principal das rotinas. Atualiza fpdf para fpdf2 em
requirements.txt devido ao uso de fpdf.enums.
