+++ requirements.txt
@@ -2,7 +2,7 @@ certifi==2025.8.3
 charset-normalizer==3.4.3
 colorama==0.4.6
 et_xmlfile==2.0.0
-fpdf==1.7.2
+fpdf2==2.8.7
 geographiclib==2.1
 geopy==2.4.1
 idna==3.10
